refactor(brave): cleanup
- move everything to separate dir
- disable debug ipfs:http-api* logs
