+++ add-on/src/lib/ipfs-client/embedded-chromesockets/config.js
