+++ add-on/src/lib/ipfs-client/embedded-chromesockets/libp2p.js
